requirements.txt
Flask==2.3.3
google-cloud-ndb==2.2.1
firebase-admin==7.0.0
Werkzeug>=2.3.7
Jinja2>=3.1.2
itsdangerous>=2.1.2
click>=8.1.3
blinker>=1.6.2